refactor: 혼란을 야기할 수 있는 패키지 명 수정 (#332)
